many: refactor store code to be able to use simpler form of auth creds (#12057)
this introduces an Authorizer interface, that can be used by the store
to authorize requests and is passed to it via config optionally.

* store: preparations, auth.go type definitions

* many: refactor store code to be able to use simpler form of auth creds

* store: some clarifications in the refactored autho code

thanks @mardy
